Add respective component folders
